marvel-dex-backend/requirements.txt
alembic
fastapi
psycopg2
pydantic
python-decouple
python-multipart
sqlmodel
uvicorn
pyjwt
passlib[bcrypt]
python-dotenv